# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/sse/connect": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "14.SSE连接"
                ],
                "summary": "建立SSE连接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token (可选，优先使用Authorization Header)",
                        "name": "token",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE Stream",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/{ids}": {
            "delete": {
                "tags": [
                    "02.用户接口"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID列表",
                        "name": "ids",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/users/{userId}": {
            "put": {
                "tags": [
                    "02.用户接口"
                ],
                "summary": "更新用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/users/{userId}/form": {
            "get": {
                "tags": [
                    "02.用户接口"
                ],
                "summary": "获取用户表单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/users/{userId}/password/reset": {
            "put": {
                "tags": [
                    "02.用户接口"
                ],
                "summary": "重置用户密码",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/users/{userId}/status": {
            "patch": {
                "tags": [
                    "02.用户接口"
                ],
                "summary": "修改用户状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
...
}